refactor: relocate Spotify API helpers
