+++ get_user_profile/lib/spotify/player.ts
